requirements.txt
# Core Dependencies
requests>=2.31.0
pyyaml>=6.0.1
openai>=1.12.0

# Optional: For future enhancements
# beautifulsoup4>=4.12.0
# lxml>=4.9.0
